+++ src/Assert/InstanceOfClass.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Nandan108\DtoToolkit\Validate;
+namespace Nandan108\DtoToolkit\Assert;
 
-use Nandan108\DtoToolkit\Core\ValidateBase;
+use Nandan108\DtoToolkit\Core\ValidatorBase;
 use Nandan108\DtoToolkit\Exception\Config\InvalidConfigException;
 use Nandan108\DtoToolkit\Exception\Process\GuardException;
 
 #[\Attribute(\Attribute::TARGET_PROPERTY | \Attribute::IS_REPEATABLE)]
-final class InstanceOfClass extends ValidateBase
+final class InstanceOfClass extends ValidatorBase
 {
     /**
      * @psalm-suppress PossiblyUnusedMethod