+++ src/Assert/NotBlank.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Nandan108\DtoToolkit\Validate;
+namespace Nandan108\DtoToolkit\Assert;
 
-use Nandan108\DtoToolkit\Core\ValidateBase;
+use Nandan108\DtoToolkit\Core\ValidatorBase;
 use Nandan108\DtoToolkit\Exception\Process\GuardException;
 
 #[\Attribute(\Attribute::TARGET_PROPERTY | \Attribute::IS_REPEATABLE)]
-final class NotBlank extends ValidateBase
+final class NotBlank extends ValidatorBase
 {
     /** @psalm-suppress PossiblyUnusedMethod */
     public function __construct(bool $trim = true)