+++ src/Assert/Regex.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Nandan108\DtoToolkit\Validate;
+namespace Nandan108\DtoToolkit\Assert;
 
-use Nandan108\DtoToolkit\Core\ValidateBase;
+use Nandan108\DtoToolkit\Core\ValidatorBase;
 use Nandan108\DtoToolkit\Exception\Config\InvalidArgumentException;
 use Nandan108\DtoToolkit\Exception\Process\GuardException;
 
 // use Nandan108\DtoToolkit\Exception\Config\InvalidArgumentException;
 
 #[\Attribute(\Attribute::TARGET_PROPERTY | \Attribute::IS_REPEATABLE)]
-final class Regex extends ValidateBase
+final class Regex extends ValidatorBase
 {
     /** @psalm-suppress PossiblyUnusedMethod
      *